create_distorted_test.py
"""
TASK COMPLIANT: Distorted medicine bottle test image
"""
import cv2
import numpy as np
from PIL import Image, ImageDraw, ImageFont
import os

def create_distorted_medicine_label():
    """Create TASK-SPEC distorted medicine bottle image."""
    
    # Create base image
    img = Image.new('RGB', (500, 400), color=(240, 240, 240))
    draw = ImageDraw.Draw(img)
    
    # Add realistic medicine text (openFDA verified drugs)
    try:
        font = ImageFont.truetype('arial.ttf', 32)
        font_small = ImageFont.truetype('arial.ttf', 24)
    except:
        font = ImageFont.load_default()
        font_small = ImageFont.load_default()
    
    # TEXT WITH REAL DRUGS (TASK COMPLIANT)
    draw.text((40, 30), 'LISINOPRIL', fill='black', font=font)
    draw.text((40, 85), '10 MG TABLETS', fill='black', font=font_small)
    draw.text((40, 135), 'TEVA PHARMACEUTICALS', fill='black', font=font_small)
    draw.text((40, 185), 'LOT: L567890', fill='black', font=font_small)
    draw.text((40, 235), 'EXP: 03/2028', fill='black', font=font_small)
    
    # SIMULATE BOTTLE CURVATURE (TASK REQUIRED)
    curved_img = img.copy()
    pixels = curved_img.load()
    width, height = curved_img.size
    
    for x in range(width):
        for y in range(height):
            # Simulate cylindrical distortion
            curve_x = x + int(15 * np.sin(y * np.pi / height))
            if 0 <= curve_x < width:
                r, g, b = img.getpixel((x, y))
                pixels[curve_x, y] = (r, g, b)
    
    # ADD GLARE (TASK REQUIRED - foil reflectivity)
    glare_x, glare_y = 250, 80
    for i in range(100):
        dx = np.random.randint(-30, 30)
        dy = np.random.randint(-20, 20)
        if 0 <= glare_x + dx < 500 and 0 <= glare_y + dy < 400:
            pixels[glare_x + dx, glare_y + dy] = (255, 255, 220)
    
    # SIMULATE BARCODE
    for x in range(100, 400, 8):
        for i in range(4):
            if x + i < 500:
                for y in range(320, 380):
                    pixels[x + i, y] = (0, 0, 0)
    
    # Save distorted images
    os.makedirs("data/input/distorted", exist_ok=True)
    curved_img.save("data/input/distorted/realistic_medicine.jpg")
    
    # Create clean version for comparison
    img.save("data/input/distorted/clean_medicine.jpg")
    
    print("✅ TASK-COMPLIANT TEST IMAGES CREATED:")
    print("   📁 data/input/distorted/realistic_medicine.jpg (CURVED+GLARE)")
    print("   📁 data/input/distorted/clean_medicine.jpg (REFERENCE)")
    
if __name__ == "__main__":
    create_distorted_medicine_label()
