requirements.txt
numpy
opencv-python
pillow
torch
torchvision
ultralytics
easyocr
pytesseract
pyzbar
python-dotenv
requests
fuzzywuzzy
python-Levenshtein
scikit-image
scipy
pandas
tqdm
roboflow
datasets
transformers
pydantic
aiohttp
asyncio
pytest
matplotlib
seaborn
